WIP: Bug 571722: Add ALL_FLAGS for scanner discovery
This change adds the ALL_FLAGS that does not limit tool options to
those declared as IOption::isForScannerDiscovery when launching the
compiler to discover compiler built-ins.

This is needed as many other flags, either entered manually in "Other
flags" or some of the existing flags with checkboxes such as "-ansi",
"-fPIC", and "-fstack-protector-all" which all affect scanner discovery
as they can all change what macros are built-in to the compiler.

The current solution has as a drawback that some settings, like -I and -D
then appear twice. For example in the "Includes" node in the "Project
Explorer"

TODO for this change:
- Handle indexer updating
    see ToolSettingsTab.isIndexerAffected

- Handle notifying listeners of an option change that may affect them
    see ManagedBuildManager.notifyListeners

- Changing default to ALL_FLAGS
    see implementations of org.eclipse.cdt.core.LanguageSettingsProvider
    extension point (such as in org.eclipse.cdt.managedbuilder.core/plugin.xml)

- Updating documentation
    see doc/org.eclipse.cdt.doc.user/reference/cdt_u_pref_build_scanner_discovery.htm

- N&N entry
    see https://wiki.eclipse.org/CDT/User/NewInCDT

Change-Id: Ie95fa877fb2685c6da49fcd8db3521fe49c3944f