Bug 572581: fix async behaviour in test
The test for ThreadStackFrameSyncTest accessed some state in the
IGDBFocusSynchronizer from outside the Executor thread.

The test also was not waiting on events to synchronize with
the executor thread, instead using Thread.sleep().

There was also some unneeded code due to not using the ServiceEventWaitor
in all places it could be used.

This was leading to race conditions and flaky test failures.

Change-Id: Ib6a3273e20f4383714edc491de9dd9330b250487