Update patch set 6
Patch Set 6:

(1 comment)

Okay, the virtual image trick worked well with a "hacked" formatWinErrorCode where I prepended "MAKELANGID(LANG_KOREAN, SUBLANG_DEFAULT)," to the array langId and with the Korean language pack installed in Windows 10.

The only remaining part of this gerrit is if the format is good or if the format should be different to make it more understandable.
Any preferences @Jonah?

Patch-set: 6