Update patch set 7
Patch Set 7: Code-Review-1

> Patch Set 7: Code-Review+2
> 
> @John - I would like this to land in CDT 2021-06 M1 which means I need to merge this by Monday. If you don't have a chance to review before then, that is ok and I will adapt to your review comments after M1 is out.

I fetched patchset 7 and tested the new UI. I am happy with separating the build command and the GNU make-specific build arguments in this way, but hiding the "Build settings" group on the "Behavior" page based on the "Use default build arguments" checkbox on the "Builder Settings" page is not good for usability. If the "Build settings" group is hidden, the user has no clue how to reveal it. Ideally, the "Build settings" group would be on the same page as the "Use default build arguments" checkbox and would be disabled rather than hidden, so the user would see the consequence of checking/unchecking the "Use default build arguments" checkbox immediately. If this is urgent for 2021-06 M1, perhaps we could just disable the "Build settings" group rather than hide it, and add a tooltip to refer the user to the checkbox on the "Builder Settings" page?

Patch-set: 7
Reviewer: Gerrit User 846 <846@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1