Update patch set 10
Patch Set 10:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > This is a preview of what I am working on - https://imgur.com/JQoL1Nx.png
> > 
> > with the main tab being https://imgur.com/5YULLdc.png
> > 
> > WDYT?
> 
> main tab is actually this https://imgur.com/gDzbqJe.png (only difference is the string with the mention of the Behavior tab). I tried to make the "Behavior" in the text a link to the behavior tab.

Looks better to have the arguments placed on the same page as the UI elements that builds up the arguments.
Is the "build arguments" text fields supposed to be editable when using the UI elements above it?
I haven't looked at the code, but from the image, it looks like it's taking input when it should be disabled or at least read-only.

Patch-set: 10
Reviewer: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>