Update patch set 2
Patch Set 2:

The change looks clean, but I don't know what implication it will have for running a CDT based product.
One thing that stands out is the MangedBuildManager#invokeConverter(ManagedBuildInfo, IBuildObject, IConfigurationElement) that no longer does anything if it does not have an UI. Is that expected?
I also noticed an extra bland link in that function, but that's really minor ;)

Should the entire UpdateManagedProjectManager be marked deprecated too if it heavily depend on the UI, or do we expect it to be updated to not need an UI?

Patch-set: 2