Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> The change looks clean, but I don't know what implication it will have for running a CDT based product.

Me neither, beyond what I have written below.

> One thing that stands out is the MangedBuildManager#invokeConverter(ManagedBuildInfo, IBuildObject, IConfigurationElement) that no longer does anything if it does not have an UI. Is that expected?

That is correct - but AFAIU it is dead code. No one is using CDT projects from 15 years ago. What I don't know is if anyone is sitting on top of that to do conversions still from custom project types.

> I also noticed an extra bland link in that function, but that's really minor ;)
:-(


> 
> Should the entire UpdateManagedProjectManager be marked deprecated too if it heavily depend on the UI, or do we expect it to be updated to not need an UI?
Probably. 

What I didn't do in this commit was mark all the old code that deals with older versions of CDT projects as deprecated. I started working on that in https://bugs.eclipse.org/bugs/show_bug.cgi?id=568510 and found that the intertwining was so high that I couldn't untangle it.

Patch-set: 2