Bug 573110: Emulator can spin if Reader is ready but not available
The implementation of the read loop in the emulator can spin because
it uses Reader.ready() to determine if there is more data to read.

However the Reader contract does not specify that ready() means
that read() will return a character, simply it means that read()
won't block. As such, if a Reader won't block, but it has no
characters, the inner read loop will spin constantly polling.

The outer loop uses polling too - but it has a wait so that the
CPU does not hit 100% and yields.

Change-Id: Id9b2426c65e6c2a2c3ae817a78d2be435e568c1f