Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > Is this a good solution to the problem reported in the bugzilla?
> > We have got similar feedback from our users...
> 
> I suspect that the full solution (preserving settings for other toolchains that are not in use anymore) is too big to solve for the benefit it would give. Therefore I am +1 on at least warning users that they will lose everything (and hopefully they will have committed their changes before doing this!)

Yes, unfortunately, that a much wider problem.
If we were to keep the values, that would mean that the set of options in the .cproject file would just grow and there would be a lot of options that would not be in use. I'm not sure if that's better than purging them when selecting a different toolchain.
The removal of the old options/toolchain is at FolderInfoModification#setToolChain(IToolChain,boolean) if you want to take a look.

I've updated the message with your suggestion. If there is no more feedback by Monday afternoon (Europe time ;)), I will merge this.

Patch-set: 2