Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> (1 comment)
> 
> Hi Jonah,
> thanks for reviewing. I hope that the NPEs are handled correctly now.

Hi Marc - sorry, this went in the wrong direction, the FAQ I linked to was probably not that useful. Unfortunately the code in PS2 follows the pattern that the FAQ recommends against. I have provided a PS3 which just adds the missing null checks (compared to PS1). Please let me know (with a +1 or comment) if this change is ok with you and I will submit it.

BTW it could be that in practice that these methods never return null - but as their API advertise they can return null I added the null checks.

See this for the change I made: https://git.eclipse.org/r/c/cdt/org.eclipse.cdt/+/180162/1..3

The rest of the patch looks great. BTW did you see my Comment 4 on the bug? Sorry it came in a little late.

Patch-set: 3
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>