Update patch set 2
Patch Set 2:

Thanks Lidia for applying the review changes. 

Alexander is right - StubUtility is probably not the correct place for the new code, as it is specific to the dialog, the new methods can be moved right into WizardNewFileFromTemplateCreationPage

Apart from that this is ready to go. If you want me to just move the methods and constant into WizardNewFileFromTemplateCreationPage let me know and I can do that and merge the change.

Thanks.

Patch-set: 2