Update patch set 1
Patch Set 1:

This looks good to me. Sadly it looks like you have a bunch of generated makefiles to update to get all tests passing again. It is really useful to see the effective makefiles that this change to the code makes, but it is a pain to update so many files. If you have any ideas on lowering the effort on such changes that would be welcome.

There are a bunch of other failures which seem weird and likely unconnected??? 

  org.eclipse.core.internal.resources.ResourceException: Errors occurred while refreshing resources with the local file system.
	at org.eclipse.cdt.managedbuilder.core.tests.BuildDescriptionModelTests.tearDown(BuildDescriptionModelTests.java:873)

Patch-set: 1
CC: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>