Update patch set 1
Patch Set 1:

(1 comment)

Thanks Marc for the patch. This looks good on a first pass. When I get to my computer I will run the code to try it out. 

This note is for other reviewers: I was going to add a few additional comments to the code, then I realised that the new code followed the same style as the existing (E.g. not using try-with-resources). If someone wants to refactor and/or improve the style here, please pull that into a separate bug/discussion.

Marc, while the code is copied, I don't think we should leave the recurse == true case in the code as that introduces new dead code. WDYT about adding removeAll or removing the recurse == true branches?

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>