Bug 572880: Keep internal representation of opcode as a String
We receive a string from GDB and then display the same string to the
user. So instead of converting it internally into a Byte[] to then
convert it back into a String, keep it as a String.

This fixes a regression where some GDBs' output format was not as space
separated 8-bit-bytes, but as words or similar.

Change-Id: I4ea241ff9ea45165489604fee26a3593ec3f6756