Bug 573646: Restore to correct secondary tab
The handling of ssecondary ids was problematic, special cases were
not well documented and depending on code paths, a secondary ID
of null would mean different things.

This commit fixes that documentation, defines shared constants, and
fixes the logic so secondary id dispayed terminals restore to the
correct location on restart.

Change-Id: Ia87b64f735a37d505e1a5a7b7a4a8e210dfb8c1e