Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > https://ci.eclipse.org/cdt/job/cdt-verify-terminal-only-pipeline/45/ : ABORTED
> 
> Looks like this timed out in your new test - I have retriggered it just in case.
> 
> BTW I am not reviewing this change yet. Let me know when/if you want me to have a closer look.
> 
> To get it into 10.3 it will need to be reviewed and merged by May 31.

Hi Jonah,

thanks for the early feedback. Yes, this is very much an attempt yet (there are still issues).
I will write here when the change is in a good enough shape for review.

I'm surprised about the automatic re-flow on windows as the eclipse terminal keeps its own line wrapping information. Will have a look at it on a windows box.

Patch-set: 1