Update patch set 1
Patch Set 1:

> I'm surprised about the automatic re-flow on windows as the eclipse terminal keeps its own line wrapping information. Will have a look at it on a windows box.

I am normally on Linux, but I have been doing a bunch of work on Terminal for Windows recently. The implementation of the PTY there is massively different because the shell does not "talk" to Eclipse directly, but via an intermediary (ConPTY or WinPTY) and that intermediary does the reflowing, not the CDT code. So when the window is resized, ConPTY sends a clear screen and resends everything.

In other words, it may not be worth comparing the Windows one too much.

Patch-set: 1