WIP - Bug 573796: Inherit expected charset from the PTY
The PTY, in case of ConPTY, operates always in UTF-8. This should be
enforced back on the user.

WIP because I don't know how/where/if the user should be able to change
it anyway, and if not to show the user that their setting has not
taken effect. If on initial connect the user has selected non-UTF-8 then
the status bar will show the actual connection as UTF-8, and the user
(if they really want to?) can change the encoding and it will take effect.

Change-Id: Iaf4c13e256ea7ee3469eecab4a0ec3df1ceb19a4