Update patch set 1
Patch Set 1:

Based on the work I did for @launcher.dir, this property should be what we thought was not available earlier.
As this property is not available in a runtime, it's a bit hard to test this.

What is the proper way of building a complete CDT product and not just the update sites?

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>