Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > @Jonah, I'm not sure if setHistoryKnown(true) should be called before checking the flags or after.
> > Regardless, the MemoryByte instance for the same address should compared...
> > 
> > Do you have any idea about the history known?
> 
> I don't know about that area, sorry.

Are you aware of any unit test in CDT that relies on a predefined pattern of mi-commands?
I mean, the first time your read a certain memory address it should be fine, but the 2nd time, GDB should reply that the read "failed".
Even better would be if this would be possible to simulate on a PC, but I've not figured out how to do that yet. Do you have any idea about this kind of tests or is this too advanced for a unit test?

Patch-set: 2