Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Just a not that there are many types of pragmas, It wouldn't make sense to display all of them in the outline view, see
> https://docs.microsoft.com/en-us/cpp/preprocessor/pragma-directives-and-the-pragma-keyword?view=msvc-160

I use following pragma's frequently
once 
warning
error

Probably comment is a good one to.

Note that there is no consistency between the GCC pragma's and the microsoft pragma's but I found these 4 above in both

also note that https://gcc.gnu.org/onlinedocs/gcc/Pragmas.html states  "Note that in general we do not recommend the use of pragmas;"

The mark pragma is a specific one for xcode https://stackoverflow.com/questions/6853821/what-is-the-significance-of-pragma-marks-why-do-we-need-pragma-marks#6853831
I'm currently making the assumption build tools ignore pragma's they do not know.

Patch-set: 1
CC: Gerrit User 96955 <96955@97ee7c02-f12f-4043-b43e-dea463d88b31>