Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Just a not that there are many types of pragmas, It wouldn't make sense to display all of them in the outline view, see
> > https://docs.microsoft.com/en-us/cpp/preprocessor/pragma-directives-and-the-pragma-keyword?view=msvc-160
> 
> +1 - I agree - but when should they be filtered out, when creating the model, or when resolving in org.eclipse.cdt.internal.ui.BaseCElementContentProvider.getTranslationUnitChildren(ITranslationUnit)?

I'm not sure. I initially thought that we could create a type of element that would be more specific with a name that would both match '#pragma mark' and '#pragma region'. It would mean a bit less unnecessary elements for clients and less work on filtering. But looking at how both type of pragmas work, I realize now that '#pragma mark' works differently in that it doesn't need to be "ended" manually, like '#pragma endregion'. So perhaps it's better to have a more generic CElement type that covers all pragmas as it will leave more flexibility for different pragmas and introduces a bit less arbitrary concept at the CElement level. At that expense of a bit more work for client code.

Patch-set: 2