Update patch set 4
Patch Set 4:

If I don't miss something all you comments are summarized in PS4.

Thinking about test-suites, being not so familiar with CDT test infra, I'm not sure how to write a test-case for this. If required I may have to allocate more time to learn this part too.

Patch-set: 4