Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Marc-Andre - your explanation seems reasonable to me, but I don't have the understanding to know what impacts there may be. Getting this into M1 is good timing.

Jonah, I wanted to give you some visibility on the fact that the CDT semantic code does not handle multiple C++ versions and historically it was a best effort to not break old C++ at the same time as supporting the new versions. But this patch is an example where this is not possible, and this might affect part of the community using older C++...hopefully just slightly.

Patch-set: 2