Update patch set 15
Patch Set 15:

Hi. I can help review this and test for M1 (likely not in next 3 days though).

One thing I'd like to mention before getting into code details is that pragma region can be nested so they need to be modelled and displayed differently. For example:

```
#pragma region Test

void foo();

#pragma region Test2

void bar();

#pragma endregion

void bazz();

#pragma endregion
```

With this change, bazz will look like it's contained in Test2 but it should be in Test.

Unless you want to go through all the extra trouble to support pragma region in the same change, I would suggest removing it from this change and focus on pragma mark.

Patch-set: 15