Update patch set 2
Patch Set 2:

> OK, then go ahead!
> Another thing that is missing: Make sure the plugins for non-gcc/clang support (arm, hpenonstop, intel, ms, nvidia) show up in an 'optional' category in the plugin installer GUI.

There are two aspects to this request - 1. the install new software wizard in the IDE does not have such a mechanism, so we rely on naming of the feature and/or category. 2. The marketplace entry can certainly have this added, but that is a non-code change.

Patch-set: 2
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>