Bug 574247 - Same binary file can appear multiple times
A race condition could sometimes yield duplicate entries in
the binary container due to interleaving of calls to
includesChild() and addChild()

Use a Set instead of a List for the collection that holds the
element's children. This avoids getting duplicate entries when
multiple threads add the same element.

This avoids the need to add synchronisation further up the call
stack -- the collection containing the children is already a
synchronised collection anyway.

Change-Id: I1ef1cddf3aad4934ec63cb433ebae34a77b69739
Signed-off-by: Mat Booth <mat.booth@gmail.com>