Update patch set 3
Patch Set 3: Code-Review-2

(1 comment)

Thanks Mat for this change, but we can't change the list to a set as it breaks the model in org.eclipse.cdt.internal.core.model.CModelBuilder2.buildModel(IASTTranslationUnit) with regard to sorting.

To see an example in practice, look at the outline view with a C file like this:

#define X
void method();
#define Y
void method2();

prior to this change the outline shows it in the order of the C file, after the change the macros will be grouped at the top (because they are inserted first in buildModel).

Note that most CElements are also ISourceReference which have getIndex to differentiate otherwise equal elements. Some CElements (such as BinaryElement) do not and there can be multiple entries that are equal. I don't actually know how to demonstrate this case though.

Patch-set: 3
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2