Update patch set 4
Patch Set 4:

(1 comment)

@Jonah, how about this instead?

Adding a method to CElementInfo to do the check+add atomically which synchronised on fChildren for the duration of both operations. The Vector#contains() and Vector#add() methods also individually synchronise on the fChildren instance, but obviously we don't want to let go between the operations.

Patch-set: 4