Update patch set 4
Patch Set 4: Code-Review+1

It looks like in this case (the stack traces) that the binary would be parsed twice, which is unfortunate. I recall binary parsing was a bit of an odd ball in the CModel given how asynchronous it is. I think since this fix is nicely localized, this is the way to go. Reading initial comments without the code, I was concerns about deadlocks which happened many times in the model and related jobs but I don't see any risk here.

Patch-set: 4
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1