Update patch set 3
Patch Set 3:

(1 comment)

Is this ready for review? I'll need a bit of time to integrate this in my codebase and switch some projects to use it.
At first glance, it does seem a bit different. For example, it doesn't seem to skip parsing if cdb is unmodified (based on modified timestamp or otherwise) and there is no setting for the path. My CDBs live in another folder than the CWD of the build.

Patch-set: 3
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>