Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3:
> 
> (1 comment)
> 
> Is this ready for review? I'll need a bit of time to integrate this in my 

Yes, it is.
> codebase and switch some projects to use it.

> At first glance, it does seem a bit different. For example, it doesn't seem to skip parsing if cdb is unmodified (based on modified timestamp or otherwise) and 
there is no setting for the path.


That is handled by jsoncdb code.

> My CDBs live in another folder than the CWD of the build.

I wasn't aware of that. My code simply follows the convention of the CDB spec [1] which recommands to put the compile_commands.json in the top level build dir.
Anyway, we could also add an input field to allow different locations.


[1[ https://clang.llvm.org/docs/JSONCompilationDatabase.html#build-system-integration

Patch-set: 3