Update patch set 1
Patch Set 1:

(1 comment)

Looks correct.

In the N&N will need a comment saying if user was relying on modifying the args passed to rm (by assigning to OBJS, etc) then they have to update their makefile.defs/init to chain the cleans like you do in the subdir.mk. However I don't think there will be many users doing this - and a small modification to their clean line is a small price to pay to have this properly fixed.

A bunch of the tests failed as expected - but a bunch failed with "Errors occurred while refreshing resources with the local file system." That either indicates the clean command line changes causes some error - or there was a problem running the build. I retriggered in case it is the latter.

I made some comments on some details to have a look at once the high level issues are dealt with.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>