Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Looks correct.
> 
> In the N&N will need a comment saying if user was relying on modifying the args passed to rm (by assigning to OBJS, etc) then they have to update their makefile.defs/init to chain the cleans like you do in the subdir.mk. However I don't think there will be many users doing this - and a small modification to their clean line is a small price to pay to have this properly fixed.

Absolutely, that should be added too!

> A bunch of the tests failed as expected - but a bunch failed with "Errors occurred while refreshing resources with the local file system." That either indicates the clean command line changes causes some error - or there was a problem running the build. I retriggered in case it is the latter.
> 
> I made some comments on some details to have a look at once the high level issues are dealt with.

I've looked at both builds in Jenkins but cannot understand why the refresh issue pops up now. I've also tried to run the build locally, but that gives me a few errors relate to the baseline compare that the META-INF/ECLIPSE_.RSA and META-INF/ECLIPSE_.SF differs. Don't think I can generate those locally though...
Running the org.eclipse.cdt.managedbuilder.core.tests as a JUnit Plug-in Test produces some 400 errors on my system, while most of them are NPE or similar. Those should not be related to the clean change!

When the name pattern of the clean target for each of the subdir.mk has been decided, I will fix the failures related to the changes in subdir.mk and leave the others for a second pass if needed.

Patch-set: 1