Update patch set 2
Patch Set 2:

Is there any point in spawning the job when we know already that the list should be empty on non-Windows systems?

The changeset might look big, but the change is only to assign result in the constructor for non-Windows systems (and remove the OS check in the job itself).

@Jonah: I do not have any WSL enabled system to test on, so if you feel like it, please feel free to run a quick check to verify that it still works :)

Patch-set: 2