Bug 575508: Fix flaky test
Bug_303953Test.testBuildAfterSourcefileDelete() is occasionally seen
failing with an unexpected change to .cproject. I have not investigated
what the change is, but it's plausible that some background process like
the indexer or scanner discovery might asynchronously modify .cproject,
and .cproject is not the focus of this test, so just ignore it along
with .project.

On the hopeful assumption that this was the only flakiness, I am
removing the @Tag(BaseTestCase5.FLAKY_TEST_TAG) annotation.

Change-Id: Ie9191aaf1a66e8475f05eb422d6e4b3949f8138d
Signed-off-by: Christian Walther <walther@indel.ch>