Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> (4 comments)
> 
> I reviewed about half the testcases. But the items that aren't being deleted anymore looks like it needs to be resolved first. I will review all the test cases once we agree a way forward on that.

I created a small python script that updated all subdir.mk and makefile in the Benchmarks directories using a certain pattern as a base for reviewing the test result. As testing is so slow on that machine, I decided to push it to have Jenkins give me a status of the test so that I can manually correct those that are not according to the generic pattern.

As there are some makefiles that look a bit different, I need to debug them to identify what's the correct way to handle it.

No need to review this gerrit more until I explicitly ask you to do it :)

Patch-set: 5