Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Quick question before I dig deeper:
> 
> In GnuMakefileGenerator,
>   public HashMap<String, List<IPath>> getBuildOutputVars()
> is API - changing the internal variable to a TreeMap (not a subclass of HashMap, unlike LinkedHashMap) would probably require breaking that. Is this worth breaking API or should I look for a different solution?

If LinkedHashMap gives stability I think that should be fine - I assume that will keep the output order stable enough. It is a shame that HashMap is the return type and not Map, but that decision was made a long time ago (2005). The method returns a modifiable map too, so can't return a copy either.

If you don't think that is stable enough, lets discuss further. For example we could copy contents into a TreeMap before writing to get the order?

Patch-set: 1