Update patch set 1
Patch Set 1:

Plot twist: It’s not that my seemingly unrelated code change changed the iteration order of the HashMap. It’s that the tests use a reordering-tolerant file comparator (org.eclipse.cdt.managedbuilder.testplugin.ManagedBuildTestHelper.compareMakefiles()) that let the test succeed as long as reorderings were the only differences. Only when my change caused a removed line, the comparison failed and output a diff *including* the reordering that was not caused by my change but had been there all along.

It’s probably still worth going ahead with the LinkedHashMap. It preserves the insertion order, which as far as I examined seems to come from the toolchain definitions (not from other nondeterministic sources), which is probably good enough. I’m still figuring out the best way of finding which benchmark files should be adjusted to the new (hopefully stable) ordering so that future test failures will not generate spurious reordering diffs.

Patch-set: 1