Bug 575628 - jsoncdb parser defines normal includes as local includes
When using jsoncdb and parsing either -I (POSIX) or /I (CL) arguments,
the include paths are flagged as local include paths. Local include
paths are treated by the CDT preprocessor only for #include "foo.h"
(quoted paths). So when using jsoncdb, several of my #include <foo.h>
don't work because they cannot be found on the wrongly flagged local
include path whereas others can sometimes be found because of a feature
in the indexer preferences called "heuristic resolution" which works if
the header happens to be in the workspace (but it can also silently pick
up the wrong one on name clash).

Change-Id: Ibb5317ed73f4a1700f21e8e3531db74995542d13
Signed-off-by: Marc-Andre Laperle <malaperle@gmail.com>