Update patch set 7
Patch Set 7:

(8 comments)

I think this is good.

As for the API which is binary compatiable - but not behaviour compatible - that is sort of the nature of the evolution of this stuff. We can't break existing extenders, but we also can't be stuck forever.

Perhaps @John has some thoughts on those particular bits of API. 

embed-cdt's use doesn't seem to conflict[1], but I have only checked subclass, not uses.

[1] https://github.com/eclipse-embed-cdt/eclipse-plugins/blob/c57bac116b482ee102ebbeb875efdaea1cf25b32/plugins/org.eclipse.embedcdt.managedbuild.cross.core/src/org/eclipse/embedcdt/managedbuild/cross/core/GnuMcuMakefileGenerator.java

Patch-set: 7
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>