Update patch set 8
Patch Set 8:

(3 comments)

> Patch Set 8:
> 
> (3 comments)
> 
> There are definitely third-party plugins that extend GnuMakefileGenerator. This patch, in current form, may break such extenders in a severe way. I don't see an easy solution, but we could *consider* making the clean behaviour configurable (eg via a GnuMakefileGenerator.setPerFolderClean() method) if others agree. As it is, third-parties may question the validity of deprecating methods but rendering them immediately unusable.

The main problem (as also mentioned in the line comment) is how to decide what behavior to use. Adding a new method for it would force everyone to sub-class to enable the per directory cleaning and I assume that we want this by default to avoid the bug...?
So yes, keeping the method bodies instead of a throwing a RuntimeException is doable, but then there is absolutely no feedback that what the extender do is unsupported and does not yield any result in the generated makefile.

I guess we could simply create a copy of the class and do the change, but I would like to avoid that if possible.

Let me know how you look at this and if there is a good solution to the dilemma.

Patch-set: 8